+++ imbrium/predictors/multivarhybrid.py
@@ -626,9 +626,13 @@ class OptimizeHybridMulti(HybridMulti):
     def create_fit_cnnrnn(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        conv_block_one: int = 1,
+        conv_block_two: int = 1,
+        rnn_block_one: int = 1,
+        rnn_block_two: int = 1,
         layer_config={
             "layer0": (64, 1, "relu", 0.0, 0.0),
             "layer1": (32, 1, "relu", 0.0, 0.0),
@@ -645,9 +649,13 @@ class OptimizeHybridMulti(HybridMulti):
         """Creates CNN-RNN hybrid model."""
         self.create_cnnrnn(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            conv_block_one=conv_block_one,
+            conv_block_two=conv_block_two,
+            rnn_block_one=rnn_block_one,
+            rnn_block_two=rnn_block_two,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -662,9 +670,13 @@ class OptimizeHybridMulti(HybridMulti):
     def create_fit_cnnlstm(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        conv_block_one: int = 1,
+        conv_block_two: int = 1,
+        lstm_block_one: int = 1,
+        lstm_block_two: int = 1,
         layer_config={
             "layer0": (64, 1, "relu", 0.0, 0.0),
             "layer1": (32, 1, "relu", 0.0, 0.0),
@@ -681,9 +693,13 @@ class OptimizeHybridMulti(HybridMulti):
         """Creates CNN-LSTM hybrid model."""
         self.create_cnnlstm(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            conv_block_one=conv_block_one,
+            conv_block_two=conv_block_two,
+            lstm_block_one=lstm_block_one,
+            lstm_block_two=lstm_block_two,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -698,9 +714,13 @@ class OptimizeHybridMulti(HybridMulti):
     def create_fit_cnngru(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        conv_block_one: int = 1,
+        conv_block_two: int = 1,
+        gru_block_one: int = 1,
+        gru_block_two: int = 1,
         layer_config={
             "layer0": (64, 1, "relu", 0.0, 0.0),
             "layer1": (32, 1, "relu", 0.0, 0.0),
@@ -717,9 +737,13 @@ class OptimizeHybridMulti(HybridMulti):
         """Creates CNN-GRU hybrid model."""
         self.create_cnngru(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            conv_block_one=conv_block_one,
+            conv_block_two=conv_block_two,
+            gru_block_one=gru_block_one,
+            gru_block_two=gru_block_two,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -734,9 +758,13 @@ class OptimizeHybridMulti(HybridMulti):
     def create_fit_cnnbirnn(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        conv_block_one: int = 1,
+        conv_block_two: int = 1,
+        birnn_block_one: int = 1,
+        rnn_block_one: int = 1,
         layer_config={
             "layer0": (64, 1, "relu", 0.0, 0.0),
             "layer1": (32, 1, "relu", 0.0, 0.0),
@@ -753,9 +781,13 @@ class OptimizeHybridMulti(HybridMulti):
         """Creates CNN-BiRNN hybrid model."""
         self.create_cnnbirnn(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            conv_block_one=conv_block_one,
+            conv_block_two=conv_block_two,
+            birnn_block_one=birnn_block_one,
+            rnn_block_one=rnn_block_one,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -770,9 +802,13 @@ class OptimizeHybridMulti(HybridMulti):
     def create_fit_cnnbilstm(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        conv_block_one: int = 1,
+        conv_block_two: int = 1,
+        bilstm_block_one: int = 1,
+        lstm_block_one: int = 1,
         layer_config={
             "layer0": (64, 1, "relu", 0.0, 0.0),
             "layer1": (32, 1, "relu", 0.0, 0.0),
@@ -789,9 +825,13 @@ class OptimizeHybridMulti(HybridMulti):
         """Creates CNN-BiLSTM hybrid model."""
         self.create_cnnbilstm(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            conv_block_one=conv_block_one,
+            conv_block_two=conv_block_two,
+            bilstm_block_one=bilstm_block_one,
+            lstm_block_one=lstm_block_one,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -806,9 +846,13 @@ class OptimizeHybridMulti(HybridMulti):
     def create_fit_cnnbigru(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        conv_block_one: int = 1,
+        conv_block_two: int = 1,
+        bigru_block_one: int = 1,
+        gru_block_one: int = 1,
         layer_config={
             "layer0": (64, 1, "relu", 0.0, 0.0),
             "layer1": (32, 1, "relu", 0.0, 0.0),
@@ -825,9 +869,13 @@ class OptimizeHybridMulti(HybridMulti):
         """Creates CNN-BiGRU hybrid model."""
         self.create_cnnbigru(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            conv_block_one=conv_block_one,
+            conv_block_two=conv_block_two,
+            bigru_block_one=bigru_block_one,
+            gru_block_one=gru_block_one,
             layer_config=layer_config,
         )
         self.fit_model(