+++ imbrium/predictors/multivarpure.py
@@ -949,9 +949,12 @@ class OptimizePureMulti(PureMulti):
     def create_fit_mlp(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        dense_block_one: int = 1,
+        dense_block_two: int = 1,
+        dense_block_three: int = 1,
         layer_config: dict = {
             "layer0": (
                 50,
@@ -976,9 +979,12 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a Multi-Layer-Perceptron model."""
         self.create_mlp(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            dense_block_one=dense_block_one,
+            dense_block_two=dense_block_two,
+            dense_block_three=dense_block_three,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -993,9 +999,12 @@ class OptimizePureMulti(PureMulti):
     def create_fit_rnn(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        rnn_block_one: int = 1,
+        rnn_block_two: int = 1,
+        rnn_block_three: int = 1,
         layer_config: dict = {
             "layer0": (
                 40,
@@ -1020,9 +1029,12 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a RNN model."""
         self.create_rnn(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            rnn_block_one=rnn_block_one,
+            rnn_block_two=rnn_block_two,
+            rnn_block_three=rnn_block_three,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1037,9 +1049,12 @@ class OptimizePureMulti(PureMulti):
     def create_fit_lstm(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        lstm_block_one: int = 1,
+        lstm_block_two: int = 1,
+        lstm_block_three: int = 1,
         layer_config: dict = {
             "layer0": (
                 40,
@@ -1064,9 +1079,12 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a LSTM model."""
         self.create_lstm(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            lstm_block_one=lstm_block_one,
+            lstm_block_two=lstm_block_two,
+            lstm_block_three=lstm_block_three,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1081,9 +1099,12 @@ class OptimizePureMulti(PureMulti):
     def create_fit_cnn(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        conv_block_one: int = 1,
+        conv_block_two: int = 1,
+        dense_block_one: int = 1,
         layer_config: dict = {
             "layer0": (
                 64,
@@ -1111,9 +1132,12 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a CNN model."""
         self.create_cnn(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            conv_block_one=conv_block_one,
+            conv_block_two=conv_block_two,
+            dense_block_one=dense_block_one,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1128,9 +1152,12 @@ class OptimizePureMulti(PureMulti):
     def create_fit_gru(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        gru_block_one: int = 1,
+        gru_block_two: int = 1,
+        gru_block_three: int = 1,
         layer_config: dict = {
             "layer0": (
                 40,
@@ -1155,9 +1182,12 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a GRU model."""
         self.create_gru(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            gru_block_one=gru_block_one,
+            gru_block_two=gru_block_two,
+            gru_block_three=gru_block_three,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1172,9 +1202,11 @@ class OptimizePureMulti(PureMulti):
     def create_fit_birnn(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        birnn_block_one: int = 1,
+        rnn_block_one: int = 1,
         layer_config: dict = {
             "layer0": (50, "relu", 0.0, 0.0),
             "layer1": (50, "relu", 0.0),
@@ -1188,9 +1220,11 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a BI-RNN model."""
         self.create_birnn(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            birnn_block_one=birnn_block_one,
+            rnn_block_one=rnn_block_one,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1205,9 +1239,11 @@ class OptimizePureMulti(PureMulti):
     def create_fit_bilstm(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        bilstm_block_one: int = 1,
+        lstm_block_one: int = 1,
         layer_config: dict = {
             "layer0": (50, "relu", 0.0, 0.0),
             "layer1": (50, "relu", 0.0),
@@ -1221,9 +1257,11 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a BI-LSTM model."""
         self.create_bilstm(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            bilstm_block_one=bilstm_block_one,
+            lstm_block_one=lstm_block_one,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1238,9 +1276,11 @@ class OptimizePureMulti(PureMulti):
     def create_fit_bigru(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        bigru_block_one: int = 1,
+        gru_block_one: int = 1,
         layer_config: dict = {
             "layer0": (50, "relu", 0.0, 0.0),
             "layer1": (50, "relu", 0.0),
@@ -1254,9 +1294,11 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains a BI-GRU model."""
         self.create_bigru(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            bigru_block_one=bigru_block_one,
+            gru_block_one=gru_block_one,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1271,9 +1313,13 @@ class OptimizePureMulti(PureMulti):
     def create_fit_encdec_rnn(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        enc_rnn_block_one: int = 1,
+        enc_rnn_block_two: int = 1,
+        dec_rnn_block_one: int = 1,
+        dec_rnn_block_two: int = 1,
         layer_config: dict = {
             "layer0": (
                 100,
@@ -1304,9 +1350,13 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains an encoder-decoder RNN model."""
         self.create_encdec_rnn(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            enc_rnn_block_one=enc_rnn_block_one,
+            enc_rnn_block_two=enc_rnn_block_two,
+            dec_rnn_block_one=dec_rnn_block_one,
+            dec_rnn_block_two=dec_rnn_block_two,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1321,9 +1371,13 @@ class OptimizePureMulti(PureMulti):
     def create_fit_encdec_lstm(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        enc_lstm_block_one: int = 1,
+        enc_lstm_block_two: int = 1,
+        dec_lstm_block_one: int = 1,
+        dec_lstm_block_two: int = 1,
         layer_config: dict = {
             "layer0": (
                 100,
@@ -1354,9 +1408,13 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains an encoder-decoder LSTM model."""
         self.create_encdec_lstm(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            enc_lstm_block_one=enc_lstm_block_one,
+            enc_lstm_block_two=enc_lstm_block_two,
+            dec_lstm_block_one=dec_lstm_block_one,
+            dec_lstm_block_two=dec_lstm_block_two,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1371,9 +1429,13 @@ class OptimizePureMulti(PureMulti):
     def create_fit_encdec_gru(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        enc_gru_block_one: int = 1,
+        enc_gru_block_two: int = 1,
+        dec_gru_block_one: int = 1,
+        dec_gru_block_two: int = 1,
         layer_config: dict = {
             "layer0": (100, "relu", 0.0, 0.0),
             "layer1": (50, "relu", 0.0, 0.0),
@@ -1389,9 +1451,13 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains an encoder-decoder GRU model."""
         self.create_encdec_gru(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            enc_gru_block_one=enc_gru_block_one,
+            enc_gru_block_two=enc_gru_block_two,
+            dec_gru_block_one=dec_gru_block_one,
+            dec_gru_block_two=dec_gru_block_two,
             layer_config=layer_config,
         )
         self.fit_model(
@@ -1406,9 +1472,13 @@ class OptimizePureMulti(PureMulti):
     def create_fit_encdec_cnn(
         self,
         optimizer: str = "adam",
+        optimizer_args: dict = None,
         loss: str = "mean_squared_error",
         metrics: str = "mean_squared_error",
-        optimizer_args: dict = None,
+        enc_conv_block_one: int = 1,
+        enc_conv_block_two: int = 1,
+        dec_gru_block_one: int = 1,
+        dec_gru_block_two: int = 1,
         layer_config: dict = {
             "layer0": (64, 1, "relu", 0.0, 0.0),
             "layer1": (32, 1, "relu", 0.0, 0.0),
@@ -1425,9 +1495,13 @@ class OptimizePureMulti(PureMulti):
         """Creates and trains an encoder-decoder CNN model."""
         self.create_encdec_cnn(
             optimizer=optimizer,
+            optimizer_args=optimizer_args,
             loss=loss,
             metrics=metrics,
-            optimizer_args=optimizer_args,
+            enc_conv_block_one=enc_conv_block_one,
+            enc_conv_block_two=enc_conv_block_two,
+            dec_gru_block_one=dec_gru_block_one,
+            dec_gru_block_two=dec_gru_block_two,
             layer_config=layer_config,
         )
         self.fit_model(